+++ backend/users/views.py
@@ -1,3 +1,4 @@
+from django.shortcuts import get_object_or_404
 from djoser.views import UserViewSet
 from rest_framework import permissions, status
 from rest_framework.decorators import action
@@ -5,16 +6,15 @@ from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.response import Response
 
 from .models import Subscribe, User
-from .serializers import SubscribeSerializer, UserSerializer
-
-
-# from rest_framework.permissions import IsAuthenticated
+from .serializers import (
+    SubscribeSerializer,
+    SubscriptionSerializer,
+    UserSerializer,
+)
 
 
 class UserViewSet(UserViewSet):
-    # http_method_names = ["get", "post", "delete"]
     pagination_class = LimitOffsetPagination
-    # permission_classes = [permissions.IsAuthenticated]
     queryset = User.objects.all()
     serializer_class = UserSerializer
 
@@ -28,8 +28,16 @@ class UserViewSet(UserViewSet):
         methods=("GET",),
         permission_classes=[permissions.IsAuthenticated],
     )
-    def subscriptions(self, request, id):
-        pass
+    def subscriptions(self, request):
+        current_user = get_object_or_404(User, username=request.user)
+        subscribtions_list = current_user.user.all()
+        page = self.paginate_queryset(subscribtions_list)
+        serializer = SubscriptionSerializer(
+            page,
+            many=True,
+            context={"request": request},
+        )
+        return self.get_paginated_response(serializer.data)
 
     @action(
         detail=True,
@@ -49,10 +57,9 @@ class UserViewSet(UserViewSet):
             )
         if not check_user.exists():
             return Response(
-                {"detail": "Нельзя подписатся на не существующего автора"},
+                {"detail": "Нельзя подписатся на несуществующего автора"},
                 status=status.HTTP_404_NOT_FOUND,
             )
-        current_user = self.request.user
         serializer = SubscribeSerializer(
             data={"user": current_user.id, "subscriber": id},
             context={"request": request},
@@ -63,10 +70,18 @@ class UserViewSet(UserViewSet):
 
     @subscribe.mapping.delete
     def delete_subscribe(self, request, id):
+        check_user = User.objects.filter(id=id)
         current_user = self.request.user
         check_subscribe = Subscribe.objects.filter(
             user=current_user, subscriber=id
         )
-        if check_subscribe.exists():
-            check_subscribe.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+        if not check_user.exists():
+            return Response(
+                {"detail": "Нельзя отписатся от несуществующего автора"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+        if check_user.exists():
+            if check_subscribe.exists():
+                check_subscribe.delete()
+                return Response({"detail": "Вы отписались от автора"}, status=status.HTTP_204_NO_CONTENT)
+            return  Response({"detail": "Вы уже отписаны от автора"}, status=status.HTTP_400_BAD_REQUEST)
